requirements.txt
Django==4.2.10
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
requests==2.31.0
redis==5.0.1
django-redis==5.4.0
drf-yasg==1.21.7ls